uint16_t is actually any_castable to just int. Damn any_cast!

+++ lib/link.cpp
@@ -79,7 +79,7 @@ link_host_state::init_link(settings_provider* settings, uint16_t default_port)
     if (settings) {
         auto s_port = settings->get("port");
         if (!s_port.empty()) {
-            int port = boost::any_cast<long long>(s_port);
+            int port = boost::any_cast<int>(s_port);
             if (port > 0 && port <= 65535) {
                 default_port = port;
             }
